refactor: move PreferenceExtensions.kt to widgets and move ToolbarExtensions.kt to views
