+++ app/src/main/java/com/f0x1d/logfox/ui/fragment/settings/SettingsUIFragment.kt
@@ -6,10 +6,10 @@ import androidx.preference.Preference
 import com.f0x1d.logfox.R
 import com.f0x1d.logfox.extensions.applyTheme
 import com.f0x1d.logfox.extensions.catchingNotNumber
-import com.f0x1d.logfox.extensions.views.observeAndUpdateSummary
-import com.f0x1d.logfox.extensions.views.observeAndUpdateSummaryForList
-import com.f0x1d.logfox.extensions.views.setupAsEditTextPreference
-import com.f0x1d.logfox.extensions.views.setupAsListPreference
+import com.f0x1d.logfox.extensions.widgets.observeAndUpdateSummary
+import com.f0x1d.logfox.extensions.widgets.observeAndUpdateSummaryForList
+import com.f0x1d.logfox.extensions.widgets.setupAsEditTextPreference
+import com.f0x1d.logfox.extensions.widgets.setupAsListPreference
 import com.f0x1d.logfox.ui.fragment.settings.base.BasePreferenceFragment
 import com.f0x1d.logfox.utils.fillWithStrings
 import com.f0x1d.logfox.utils.preferences.AppPreferences
